+++ application/src/bin/curl-inject-opt.rs
@@ -4,7 +4,7 @@ use std::path::PathBuf;
 use yansi::Paint;
 
 fn main() {
-	if !curl_inject_opt::should_color(2) {
+	if !curl_inject_opt::should_color(&std::io::stderr()) {
 		Paint::disable();
 	}